fix manifest.json

+++ manifest.js
@@ -13,9 +13,11 @@ function ManifestPlugin(options) {
 ManifestPlugin.prototype.apply = function ManifestPluginApply(compiler) {
     compiler.plugin('done', () => {
         /* eslint import/no-dynamic-require: "off" */
-        let manifest = require(this.options.path);
-        manifest = Object.assign({}, manifest, this.options.replace);
-        fs.writeFileSync(this.options.path, JSON.stringify(manifest, null, 4));
+        if (fs.existsSync(this.options.path)) {
+            let manifest = require(filename);
+            manifest = Object.assign({}, manifest, this.options.replace);
+            fs.writeFileSync(this.options.path, JSON.stringify(manifest, null, 4));
+        }
     });
 };
